Make SSO work for Google Apps For Work
When trying to make SSO work for Google Apps For Work, I faced the
following error message:
  Error: invalid_request
  Invalid response_type: authorization_code
  Learn more
  That’s all we know.

The only way I made it work was to change a few parameters:
responseType, scope.
I also hade to modify the url_for() call as my securitymonkey
install is behind a SSL proxy.
There are certainly better ways to make it work but I'm not
a Flask expert and could not find a way to have it create
the proper urls.