Limit Google SSO to specific domain
In order to limit Google SSO to specific Hosted Domains, we
introduce a new configuration setting (optional):
 GOOGLE_HOSTED_DOMAIN

If GOOGLE_HOSTED_DOMAIN is not set, nothing happens.
If GOOGLE_HOSTED_DOMAIN is set in the config file, its value
is used as the parameter 'hd' in calls to Google SSO authentication
request (to GOOGLE_AUTH_ENDPOINT), and the returned ID token will
also be verified to conform to this Hosted Domain.

That way you can limit login only to users of your organisation.